[REFOCTOR]Refoctor class mainEventContorler
-this 하나때문에 3시간을 날려버렸다.
- 동료의 추천으로 Code Spell Checker 바로 설치하였다. 효과는 나중에
- 현재 header 이벤트 완벽히 구현
