+++ airbnbclone/views/index.ejs
@@ -465,7 +465,11 @@
             </div>
         </section>
     </footer>
-    <script src="./javascripts/main.js"></script>
+    <script type="module" src="./javascripts/utils.js"></script>
+    <script type="module" src="./javascripts/EventHandler.js"></script>
+    <script type="module" src="./javascripts/MainEventControler.js"></script>
+    <script type="module" src="./javascripts/HeaderEventControler.js"></script>
+    <script type="module" src="./javascripts/main.js"></script>
 </body>
 
 </html>
